Reorganize demand ware connector
