+++ web/app/integration-manager/_demandware-connector/constants.js
@@ -1,0 +1,5 @@
+const API_TYPE = 'shop'
+const API_VERSION = 'v17_2'
+export const SITE_ID = 'Sites-2017refresh-Site'
+export const DW_CLIENT_ID = '5640cc6b-f5e9-466e-9134-9853e9f9db93'
+export const API_END_POINT_URL = `/s/${SITE_ID}/dw/${API_TYPE}/${API_VERSION}`
