+++ web/app/main.jsx
@@ -19,8 +19,8 @@ import Stylesheet from './stylesheet.scss' // eslint-disable-line no-unused-vars
 import {analyticManager} from 'progressive-web-sdk/dist/analytics/analytic-manager'
 import {clientAnalytics} from './utils/analytics/client-analytics'
 
-import connector from './integration-manager/_merlins-connector'
-// import connector from './integration-manager/_demandware-connector'
+// import connector from './integration-manager/_merlins-connector'
+import connector from './integration-manager/_demandware-connector'
 
 import {registerConnector} from './integration-manager'
